refactor: migrate to Modular Monolith (Phase 1) with security hardening
Structure:
- Reorganize from layered to module-based: src/modules/{auth,user,upload}
- Extract shared code to src/shared/{constants,exceptions,interfaces,middlewares,utils}
- Add UserRepository pattern to decouple DB queries from service layer

Bug fixes:
- AuthMiddleware: add missing return after 401 response, pass errors via next()
  instead of throwing; store decoded JWT payload in req.user
- changePassword: read userId from req.user (JWT) instead of req.body
  to prevent horizontal privilege escalation
- refreshToken controller: fix incorrect env variable (ENV_DEPLOY → NODE_ENV)
- ResetPasswordDto: fix field name mismatch (password → newPassword) that
  caused reset-password endpoint to always receive undefined
- initializeMinio: move from unawaited constructor call to server.ts bootstrap
  so bucket is ready before first request

Security:
- Add authRateLimiter (10 req/15min) to register, login, forgot/reset-password
- Add uploadRateLimiter (20 req/min) to upload endpoint
- Protect GET /users with AuthMiddleware (was publicly accessible)
- Protect POST /uploads with AuthMiddleware
- Prevent user enumeration: login now returns same error for wrong username
  and wrong password ("Invalid credentials")
- Forgot-password no longer reveals if email is registered
- Add MIME type validation for uploads (JPEG/PNG/GIF/WebP only)
- Sanitize uploaded filenames to prevent path traversal
- Exclude sensitive fields (password, reset tokens) from GET /users response
- Add ChangePasswordDto validation without user_id field
- Add confirm_password field to RegisterUserDto
- Add email uniqueness check on register

Quality:
- Replace all console.log/console.error with Winston logger
- Remove debug IIFE in minio.ts that logged all bucket names on startup
