+++ package.json
@@ -35,6 +35,7 @@
     "dotenv": "^17.3.1",
     "envalid": "^8.1.1",
     "express": "^5.2.1",
+    "express-rate-limit": "^8.5.1",
     "helmet": "^8.1.0",
     "hpp": "^0.2.3",
     "i18next": "^25.8.17",
