--- src/controllers/auth.controller.ts
@@ -1,124 +1,0 @@
-import { Container } from "typedi";
-import { NODE_ENV } from "@/configs/env";
-import { IRegisterUser, IUser } from "@/interfaces/users.interface";
-import { AuthService } from "@/services/auth.service";
-import { NextFunction, Request, Response } from "express";
-export class AuthController {
-  private auth: AuthService;
-  constructor() {
-    this.auth = Container.get(AuthService);
-  }
-  // [POST]/register
-  public register = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const userData: IRegisterUser = req.body;
-      await this.auth.registerService(userData);
-      res.status(201).json({ status: 201, message: "successfully!" });
-    } catch (error) {
-      next(error);
-    }
-  };
-
-  // [POST]/Login
-  public login = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const userData: IUser = req.body;
-      const { userAvailable, accessToken, refreshToken } = await this.auth.loginService(userData);
-      const isProduction = NODE_ENV === "production";
-      // Set cookie HttpOnly
-      res.cookie("accessToken", accessToken, {
-        httpOnly: true,
-        secure: isProduction,
-        sameSite: "strict",
-        path: "/",
-        maxAge: 15 * 60 * 1000, // 15 minutes
-      });
-      res.cookie("refreshToken", refreshToken, {
-        httpOnly: true,
-        secure: isProduction,
-        sameSite: "strict",
-        path: "/",
-        maxAge: 7 * 24 * 60 * 60 * 1000, // 7 days
-      });
-      res.cookie("isLoggedIn", true, {
-        path: "/",
-        maxAge: 7 * 24 * 60 * 60 * 1000, // 7 days
-      });
-
-      res.status(200).json({
-        userId: userAvailable.id,
-        userRole: "user",
-      });
-    } catch (error) {
-      next(error);
-    }
-  };
-
-  //[POST]/Log out
-  public logout = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      res.clearCookie("accessToken", { path: "/" });
-      res.clearCookie("refreshToken", { path: "/" });
-      res.clearCookie("isLoggedIn", { path: "/" });
-      res.status(200).json({ status: 200, message: "Logged out successfully." });
-    } catch (error) {
-      next(error);
-    }
-  };
-
-  //[POST]/Refresh token
-  public refreshToken = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const { refreshToken } = req.cookies;
-      if (!refreshToken) {
-        return res.status(401).json({ status: 401, message: "No refresh token provided" });
-      }
-      const { newAccessToken } = await this.auth.refreshTokenService(refreshToken);
-      const isProduction = process.env.ENV_DEPLOY === "production";
-      // Set cookie HttpOnly
-      res.cookie("accessToken", newAccessToken, {
-        httpOnly: true,
-        secure: isProduction,
-        sameSite: "strict",
-        path: "/",
-        maxAge: 15 * 60 * 1000, // 15 minutes
-      });
-      return res.status(200).json({ status: 200, message: "Access token refreshed" });
-    } catch (error) {
-      next(error);
-    }
-  };
-
-  //[POST]/Change password
-  public changePassword = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const { user_id, current_password, new_password } = req.body;
-      await this.auth.changePasswordService(user_id, current_password, new_password);
-      res.status(200).json({ status: 200, message: "Password changed successfully." });
-    } catch (error) {
-      next(error);
-    }
-  };
-
-  //[POST]/Forgot password
-  public forgotPassword = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const { email } = req.body;
-      await this.auth.forgotPasswordService(email);
-      res.status(200).json({ status: 200, message: "Password reset email sent successfully." });
-    } catch (error) {
-      next(error);
-    }
-  };
-
-  //[POST]/Reset password
-  public resetPassword = async (req: Request, res: Response, next: NextFunction) => {
-    try {
-      const { token, newPassword } = req.body;
-      await this.auth.resetPasswordService(token, newPassword);
-      res.status(200).json({ status: 200, message: "Password reset successfully." });
-    } catch (error) {
-      next(error);
-    }
-  };
-}
