--- src/dtos/auth/forgot-password.dto.ts
@@ -1,8 +1,0 @@
-import { IsEmail, IsNotEmpty, Length } from "class-validator";
-
-export class ForgotPasswordDto {
-  @IsEmail()
-  @IsNotEmpty()
-  @Length(9, 255)
-  public email: string;
-}
