--- src/dtos/auth/login.dto.ts
@@ -1,13 +1,0 @@
-import { IsString, IsNotEmpty, Length } from "class-validator";
-
-export class LoginDto {
-  @IsString()
-  @IsNotEmpty()
-  @Length(9, 255)
-  public user_name: string;
-
-  @IsString()
-  @IsNotEmpty()
-  @Length(9, 255)
-  public password: string;
-}
