--- src/interfaces/i18n.interface.ts
@@ -1,5 +1,0 @@
-import { Request } from "express";
-import { TFunction } from "i18next";
-export interface II18nRequest extends Request {
-  t: TFunction;
-}
