--- src/middlewares/upload.middleware.ts
@@ -1,6 +1,0 @@
-import multer from "multer";
-
-export const upload = multer({
-  storage: multer.memoryStorage(),
-  limits: { fileSize: 5 * 1024 * 1024 }, // 5 MB avatar
-});
