+++ src/modules/auth/dtos/change-password.dto.ts
@@ -1,0 +1,13 @@
+import { IsString, IsNotEmpty, Length } from "class-validator";
+
+export class ChangePasswordDto {
+  @IsString()
+  @IsNotEmpty()
+  @Length(8, 255)
+  public current_password: string;
+
+  @IsString()
+  @IsNotEmpty()
+  @Length(8, 255)
+  public new_password: string;
+}
