+++ src/modules/auth/dtos/forgot-password.dto.ts
@@ -1,0 +1,7 @@
+import { IsEmail, IsNotEmpty } from "class-validator";
+
+export class ForgotPasswordDto {
+  @IsEmail()
+  @IsNotEmpty()
+  public email: string;
+}
