+++ src/modules/auth/dtos/reset-password.dto.ts
@@ -1,0 +1,12 @@
+import { Length, IsNotEmpty, IsString } from "class-validator";
+
+export class ResetPasswordDto {
+  @IsString()
+  @IsNotEmpty()
+  public token: string;
+
+  @IsString()
+  @IsNotEmpty()
+  @Length(8, 255)
+  public newPassword: string;
+}
