+++ src/modules/user/user.service.ts
@@ -1,0 +1,11 @@
+import { Service } from "typedi";
+import { UserRepository } from "./user.repository";
+
+@Service()
+export class UserService {
+  constructor(private userRepo: UserRepository) {}
+
+  async allUserService() {
+    return this.userRepo.findAll();
+  }
+}
