+++ src/shared/constants/error.ts
