+++ src/shared/interfaces/errors.interface.ts
