+++ src/shared/interfaces/express.d.ts
@@ -1,0 +1,11 @@
+declare namespace Express {
+  interface Request {
+    user?: {
+      id: number;
+      user_name: string;
+      email: string;
+      iat?: number;
+      exp?: number;
+    };
+  }
+}
