+++ src/shared/interfaces/i18n.interface.ts
@@ -1,0 +1,6 @@
+import { Request } from "express";
+import { TFunction } from "i18next";
+
+export interface II18nRequest extends Request {
+  t: TFunction;
+}
