+++ src/shared/middlewares/upload.middleware.ts
@@ -1,0 +1,6 @@
+import multer from "multer";
+
+export const upload = multer({
+  storage: multer.memoryStorage(),
+  limits: { fileSize: 5 * 1024 * 1024 },
+});
