+++ src/shared/middlewares/validation.middleware.ts
@@ -1,0 +1,26 @@
+import { plainToInstance } from "class-transformer";
+import { validateOrReject, ValidationError } from "class-validator";
+import { NextFunction, Request, Response } from "express";
+import { HttpException } from "@/shared/exceptions/httpException";
+
+export const ValidationMiddleware = (
+  type: any,
+  skipMissingProperties = false,
+  whitelist = false,
+  forbidNonWhitelisted = false,
+) => {
+  return (req: Request, res: Response, next: NextFunction) => {
+    const dto = plainToInstance(type, req.body);
+    validateOrReject(dto, { skipMissingProperties, whitelist, forbidNonWhitelisted })
+      .then(() => {
+        req.body = dto;
+        next();
+      })
+      .catch((errors: ValidationError[]) => {
+        const message = errors
+          .map((error: ValidationError) => Object.values(error.constraints))
+          .join(", ");
+        next(new HttpException(400, message));
+      });
+  };
+};
