+++ src/shared/utils/crypto.ts
@@ -1,0 +1,9 @@
+import crypto from "crypto";
+
+export const hashToken = (token: string) => {
+  return crypto.createHash("sha256").update(token).digest("hex");
+};
+
+export const generateRandomToken = () => {
+  return crypto.randomBytes(32).toString("hex");
+};
