+++ src/shared/utils/validateEnv.ts
